factor out database_backend
