+++ exp/ingest/adapters/ledger_backend_adapter.go
@@ -3,20 +3,23 @@ package ingestadapters
 import (
 	"fmt"
 
-	"github.com/stellar/go/exp/ingest"
 	"github.com/stellar/go/exp/ingest/io"
+	"github.com/stellar/go/exp/ingest/ledgerbackend"
 )
 
+// LedgerBackendAdapter provides a convenient I/O layer above the low-level LedgerBackend implementation.
 type LedgerBackendAdapter struct {
-	backend ingest.LedgerBackend
+	backend ledgerbackend.LedgerBackend
 }
 
+// GetLatestLedgerSequence returns the most recent ledger sequence number present in the backend.
 func (lba *LedgerBackendAdapter) GetLatestLedgerSequence() (uint32, error) {
-	// TODO: placeholder
-	return 0, fmt.Errorf("Not implemented yet")
+	return lba.backend.GetLatestLedgerSequence()
 }
 
+// GetLedger returns...
 func (lba *LedgerBackendAdapter) GetLedger(uint32) (io.LedgerReader, error) {
+	// func (lba *LedgerBackendAdapter) GetLedger(uint32) (io.LedgerReadCloser, error) {
 	// TODO: placeholder
 	return nil, fmt.Errorf("not implemented yet")
 }
